Refactor: refactoring Authentication API #72
