+++ backend/controllers/userController.js
@@ -28,21 +28,6 @@ module.exports = {
         }
     },
 
-    login: async function(req, res) {
-        try {
-            const token = await userService.auth(req.body);
-            
-            res.cookie('jwt', token);
-            res.status(201).send({
-                result: 'OK',
-                token
-            });
-
-        } catch(err) {
-           res.status(err.status).send(err.message);
-        }
-    },
-
     updateUser: async function(req, res) {
         try {
             const result = await userService.update(req.params.id,req.body);
