+++ backend/services/errors/RuntimeError.js
@@ -1,6 +1,5 @@
 class RuntimeError extends Error {
     constructor(message) {
-        console.log(message);
         super(message);
         this.status = 500;
         this.name = "RuntimeError";
